+++ src/main/java/poomasi/domain/member/service/MemberService.java
@@ -7,10 +7,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import poomasi.domain.member._profile.dto.request.AddressUpdateRequest;
 import poomasi.domain.member._profile.entity.MemberProfile;
-import poomasi.domain.member.dto.request.CustomerUpdateRequest;
-import poomasi.domain.member.dto.request.FarmerUpdateRequest;
 import poomasi.domain.member.dto.response.MemberResponse;
 import poomasi.domain.member.dto.response.MemberSummaryResponse;
 import poomasi.domain.member.entity.LoginType;
@@ -18,11 +15,9 @@ import poomasi.domain.member.entity.Member;
 import poomasi.domain.member.repository.MemberRepository;
 import poomasi.domain.member.dto.request.SignupRequest;
 import poomasi.domain.member.dto.response.SignUpResponse;
-import poomasi.domain.store.entity.Store;
 import poomasi.global.error.BusinessException;
 
 import static poomasi.domain.member.entity.Role.ROLE_CUSTOMER;
-import static poomasi.domain.member.entity.Role.ROLE_FARMER;
 import static poomasi.global.error.BusinessError.*;
 
 @Service
@@ -67,68 +62,12 @@ public class MemberService {
         return members.map(MemberSummaryResponse::fromEntity);
     }
 
-    @Transactional
-    public void convertToFarmer(Member member) {
-        if (member.isFarmer()) {
-            throw new BusinessException(MEMBER_ALREADY_FARMER);
-        }
-
-        member.setAddress(null, null, null, null);
-        member.setRole(ROLE_FARMER);
-        memberRepository.save(member);
-    }
-
-    @Transactional
-    public void convertToCustomer(Long memberId) {
-        Member member = findMemberById(memberId);
-
-        if (member.isCustomer()) {
-            throw new BusinessException(MEMBER_ALREADY_CUSTOMER);
-        }
-
-        member.setRole(ROLE_CUSTOMER);
-        memberRepository.save(member);
-    }
-
     public Member findMemberById(Long memberId) {
         return memberRepository.findByIdAndDeletedAtIsNull(memberId)
                 .orElseThrow(() -> new BusinessException(MEMBER_NOT_FOUND));
     }
 
-    @Transactional
-    public Member updateCustomer(Member member, CustomerUpdateRequest customerUpdateRequest)
-    {
-        if (!member.isCustomer()) {
-            throw new BusinessException(INVALID_ROLE);
-        }
-
-        updateCommonAttributes(member, customerUpdateRequest.name(),customerUpdateRequest.email(), customerUpdateRequest.password(), customerUpdateRequest.phoneNumber());
-
-        return memberRepository.save(member);
-    }
-
-    @Transactional
-    public Member updateFarmer(Member member, FarmerUpdateRequest farmerUpdateRequest)
-    {
-        if (!member.isFarmer()) {
-            throw new BusinessException(INVALID_ROLE);
-        }
-
-        updateCommonAttributes(member, farmerUpdateRequest.name(), farmerUpdateRequest.email(), farmerUpdateRequest.password(), farmerUpdateRequest.phoneNumber());
-
-        Store store = member.getOrCreateStore();
-
-        if (farmerUpdateRequest.storeName() != null) {
-            store.setName(farmerUpdateRequest.storeName());
-        }
-        if (farmerUpdateRequest.storeAddress() != null) {
-            store.setAddress(farmerUpdateRequest.storeAddress());
-        }
-
-        return memberRepository.save(member);
-    }
-
-    private void updateCommonAttributes(Member member, String name, String email, String password, String phoneNumber) {
+    public void updateCommonAttributes(Member member, String name, String email, String password, String phoneNumber) {
         if (name != null) member.setName(name);
         if (email != null) member.setEmail(email);
         if (password != null) member.setPassword(passwordEncoder.encode(password));
@@ -137,15 +76,6 @@ public class MemberService {
         if (phoneNumber != null) {
             profile.setPhoneNumber(phoneNumber);
         }
-    }
-
-    @Transactional
-    public void updateAddress(Member member, AddressUpdateRequest request) {
-        MemberProfile profile = member.getOrCreateProfile();
-
-        profile.setAddress(request.defaultAddress(), request.addressDetail(), request.coordinateX(), request.coordinateY());
-
-        memberRepository.save(member);
     }
 
     @Transactional
